Inicio Tests y actualizacion
