+++ Programacion/ProyectoFinalITV/src/main/kotlin/dev/team/proyectofinalitv/AppMain.kt
@@ -4,7 +4,7 @@ import dev.team.proyectofinalitv.config.AppConfig
 import dev.team.proyectofinalitv.controllers.ProbandoCosillasController
 import dev.team.proyectofinalitv.di.ModuleKoin
 import dev.team.proyectofinalitv.repositories.CitaRepositoryImpl
-import dev.team.proyectofinalitv.services.database.DataBaseManager
+import dev.team.proyectofinalitv.services.database.DataBaseManagerImpl
 import javafx.application.Application
 import javafx.stage.Stage
 import org.koin.core.context.startKoin
@@ -23,7 +23,7 @@ class AppMain : Application() {
         cont.onSaveCita()
         
         val configApp = AppConfig()
-        val db = DataBaseManager(configApp)
+        val db = DataBaseManagerImpl(configApp)
 
         val repoCita = CitaRepositoryImpl(db)
