+++ Programacion/ProyectoFinalITV/src/main/kotlin/dev/team/proyectofinalitv/repositories/VehiculoRepositoryImpl.kt
@@ -21,7 +21,7 @@ class VehiculoRepositoryImpl(private val databaseManager: DataBaseManager) : Veh
         // Seleccionamos la base de datos a la que realizar las consultas
         databaseManager.selectDataBase()
 
-        val statement = databaseManager.connection?.createStatement()
+        val statement = databaseManager.createStatement()
 
         val sql =
         """
@@ -30,13 +30,13 @@ class VehiculoRepositoryImpl(private val databaseManager: DataBaseManager) : Veh
         WHERE matricula = ?
         """
 
-        val preparedStatement = databaseManager.connection?.prepareStatement(sql)
+        val preparedStatement = databaseManager.prepareStatement(sql)
         preparedStatement?.setString(1, vehiculo.marca)
         preparedStatement?.setString(2, vehiculo.modelo)
         preparedStatement?.setString(3, vehiculo.fechaMatriculacion.toString())
         preparedStatement?.setString(4, vehiculo.fechaRevision.toString())
-        preparedStatement?.setString(5, vehiculo.tipoMotor)
-        preparedStatement?.setString(6, vehiculo.tipoVehiculo)
+        preparedStatement?.setString(5, vehiculo.tipoMotor.toString())
+        preparedStatement?.setString(6, vehiculo.tipoVehiculo.toString())
         preparedStatement?.setString(7, vehiculo.dniPropietario)
         preparedStatement?.setString(8, vehiculo.matricula)
 
@@ -63,7 +63,7 @@ class VehiculoRepositoryImpl(private val databaseManager: DataBaseManager) : Veh
         // Seleccionamos la base de datos a la que realizar las consultas
         databaseManager.selectDataBase()
 
-        val statement = databaseManager.connection?.createStatement()
+        val statement = databaseManager.createStatement()
 
         val sql =
             """
@@ -71,14 +71,14 @@ class VehiculoRepositoryImpl(private val databaseManager: DataBaseManager) : Veh
         VALUES (?, ?, ?, ?, ?, ?, ?, ?)
         """
 
-        val preparedStatement = databaseManager.connection?.prepareStatement(sql)
+        val preparedStatement = databaseManager.prepareStatement(sql)
         preparedStatement?.setString(1, vehiculo.matricula)
         preparedStatement?.setString(2, vehiculo.marca)
         preparedStatement?.setString(3, vehiculo.modelo)
         preparedStatement?.setString(4, vehiculo.fechaMatriculacion.toString())
         preparedStatement?.setString(5, vehiculo.fechaRevision.toString())
-        preparedStatement?.setString(6, vehiculo.tipoMotor)
-        preparedStatement?.setString(7, vehiculo.tipoVehiculo)
+        preparedStatement?.setString(6, vehiculo.tipoMotor.toString())
+        preparedStatement?.setString(7, vehiculo.tipoVehiculo.toString())
         preparedStatement?.setString(8, vehiculo.dniPropietario)
 
         preparedStatement?.executeUpdate()
