+++ Programacion/ProyectoFinalITV/src/test/kotlin/dev/team/proyectofinalitv/repositories/CitaRepositoryImplTest.kt
@@ -1,0 +1,34 @@
+package dev.team.proyectofinalitv.repositories
+
+import org.junit.jupiter.api.Test
+
+import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.TestInstance
+
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class CitaRepositoryImplTest {
+
+    @Test
+    fun findAll() {
+    }
+
+    @Test
+    fun deleteById() {
+    }
+
+    @Test
+    fun update() {
+    }
+
+    @Test
+    fun save() {
+    }
+
+    @Test
+    fun findById() {
+    }
+
+    @Test
+    fun findByMatricula() {
+    }
+}
