+++ Programacion/ProyectoFinalITV/src/test/kotlin/dev/team/proyectofinalitv/repositories/InformeRepositoryImplTest.kt
@@ -1,0 +1,18 @@
+package dev.team.proyectofinalitv.repositories
+
+import org.junit.jupiter.api.Test
+
+import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.TestInstance
+
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class InformeRepositoryImplTest {
+
+    @Test
+    fun update() {
+    }
+
+    @Test
+    fun save() {
+    }
+}
