+++ Programacion/ProyectoFinalITV/src/test/kotlin/dev/team/proyectofinalitv/repositories/PropietarioRepositoryImplTest.kt
@@ -1,0 +1,18 @@
+package dev.team.proyectofinalitv.repositories
+
+import org.junit.jupiter.api.Test
+
+import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.TestInstance
+
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class PropietarioRepositoryImplTest {
+
+    @Test
+    fun update() {
+    }
+
+    @Test
+    fun save() {
+    }
+}
